Add functions for changing images, timer and popup
Added function to change images of player and computer
pictures when a choice is selected.

Added countdown function which also disables buttons
to prevent them from being pressed while round is being played.

Added a modal to pop up when someone reaches 5 points.

Added a restart button if the player is so bad that the computer
beats him and player gets mad.